Allow storing price with VAT as float
step=any is required for type=number input to accept floats with decimal points

+++ site/app/models/Training/Applications.php
@@ -486,7 +486,7 @@ class Applications
 		string $source,
 		?int $price = null,
 		?float $vatRate = null,
-		?int $priceVat = null,
+		?float $priceVat = null,
 		?int $discount = null,
 		?string $invoiceId = null,
 		string $paid = null,
